requirements.txt
fastapi==0.98.0
pydantic==1.10.9
PyJWT==2.7.0
PyJWT==2.7.0
pymongo==4.4.0
python-dotenv==1.0.0
python_bcrypt==0.3.2
